backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2
twilio==8.10.0
requests==2.31.0
xgboost==2.0.3
joblib==1.3.2
haversine==2.8.0
